Add files via upload
This is a SQL project on a movie rental company dataset. In this project, we will use SQL to query data from different tables which are either linked or unlinked in order to answer vital questions and derive insights. 

Below is the link to the dataset used for this project:
https://1drv.ms/u/c/1ef257e18c7dc85b/EaysomNUYENKkIxuGqFZ7qoBiF2-NDyLLNI-rcz1_WYEwQ?e=Vh0KNV